web/requirements.txt
aldryn-style==1.1.4
dj-database-url==0.5.0
Django==1.11.12
django-classy-tags==0.8.0
django-cms==3.5.2
django-filer==1.3.1
django-formtools==2.1
django-mptt==0.8.7
django-polymorphic==2.0.2
django-sekizai==0.10.0
django-taggit==0.22.2
django-taggit-autosuggest==0.3.3
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-installer==1.0.0
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
easy-thumbnails==2.5
gunicorn==19.8.1
html5lib==0.9999999
Pillow==5.1.0
psycopg2==2.7.4
pytz==2018.4
six==1.11.0
tzlocal==1.5.1
Unidecode==0.4.21
